pyproject.toml
[project]
name = "simpleserversetup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mcp[cli]==1.6.0",
    "nest-asyncio>=1.6.0",
    "openai==1.75.0",
    "python-dotenv>=1.1.0",
]
